[BoxLayout] Added PageLayout & BoxLayout fixups
* Fixed Rect to always set maxSize and BoxLayout navigation
* removed commented out focusItem
* Removed redundant handling
* General BoxLayout fixes

+++ module-gui/gui/core/BoundingBox.hpp
@@ -17,12 +17,12 @@ public:
 	int16_t x,y;
 	uint16_t w,h;
 	BoundingBox();
-	BoundingBox( int16_t x, int16_t y, uint16_t w, uint16_t h);
-	virtual ~BoundingBox();
+    BoundingBox(int32_t x, int32_t y, uint32_t w, uint32_t h);
+    virtual ~BoundingBox();
 
-	static bool intersect( const BoundingBox& box1, const BoundingBox& box2, BoundingBox& result );
+    static bool intersect(const BoundingBox &box1, const BoundingBox &box2, BoundingBox &result);
 
-	//inline
+    //inline
 	void clear() { x=0; y=0; w=0; h=0; };
 
 };
